refactor(ai-agent): extract RemoteAgentManager from mod.rs into agent.rs
mod.rs is now aggregation-only per target.md §2: pub mod + pub use.
RemoteAgentManager + RemoteState + RemoteAgentConfig + IAgentTask impl +
tests all moved to manager/remote/agent.rs intact (byte-identical bodies).

Fixes clippy::module_inception that surfaced during Stage 9 cleanup.
